Push Pull Block updated, Player Updated, Player friction fixed
Updated the push block so that it can be pulled as well, and fixed it from moving too much when the player jumps. Gave the player a trigger so that the move block can detect the player's range. Made it so if the player isn't moving and is grounded then they become kinematic and don't move so that slopes aren't an issue.